# openapi.json
{
    "openapi": "3.1.0",
    "info": {
      "title": "FastAPI",
      "version": "0.1.0"
    },
    "paths": {
...
      "/offers/": {
        "get": {
          "tags": [
            "Offers"
          ],
          "summary": "Get All Offers",
          "operationId": "get_all_offers_offers__get",
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "items": {
                      "$ref": "#/components/schemas/OfferResponse"
                    },
                    "type": "array",
                    "title": "Response Get All Offers Offers  Get"
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "Offers"
          ],
          "summary": "Create Offer",
          "operationId": "create_offer_offers__post",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfferCreate"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/OfferResponse"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/offers/{offer_id}": {
        "get": {
          "tags": [
            "Offers"
          ],
          "summary": "Get Offer By Id",
          "operationId": "get_offer_by_id_offers__offer_id__get",
          "parameters": [
            {
              "name": "offer_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "Offer Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/OfferResponse"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        },
        "put": {
          "tags": [
            "Offers"
          ],
          "summary": "Update Offer",
          "operationId": "update_offer_offers__offer_id__put",
          "parameters": [
            {
              "name": "offer_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "Offer Id"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfferCreate"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/OfferResponse"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        },
        "delete": {
          "tags": [
            "Offers"
          ],
          "summary": "Delete Offer",
          "operationId": "delete_offer_offers__offer_id__delete",
          "parameters": [
            {
              "name": "offer_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "Offer Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/OfferResponse"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
...
      "/campaigns/{campaign_id}/offers/{offer_id}": {
        "post": {
          "tags": [
            "Campaigns"
          ],
          "summary": "Link Offer To Campaign",
          "operationId": "link_offer_to_campaign_campaigns__campaign_id__offers__offer_id__post",
          "parameters": [
            {
              "name": "campaign_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "Campaign Id"
              }
            },
            {
              "name": "offer_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "Offer Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
  
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
...
        "HTTPValidationError": {
          "properties": {
            "detail": {
              "items": {
                "$ref": "#/components/schemas/ValidationError"
              },
              "type": "array",
              "title": "Detail"
            }
          },
          "type": "object",
          "title": "HTTPValidationError"
        },
        "OfferCreate": {
          "properties": {
            "offer_name": {
              "type": "string",
              "title": "Offer Name"
            },
            "offer_type": {
              "type": "string",
              "title": "Offer Type"
            },
            "cashback_percentage": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cashback Percentage"
            },
            "discount_amount": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Discount Amount"
            },
            "reward_points": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Reward Points"
            },
            "interest_rate": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Interest Rate"
            },
            "fee_waiver": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Fee Waiver"
            },
            "miles_earned": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Miles Earned"
            },
            "category": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            },
            "valid_from": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Valid From"
            },
            "valid_to": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Valid To"
            },
            "status": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          "type": "object",
          "required": [
            "offer_name",
            "offer_type",
            "cashback_percentage",
            "discount_amount",
            "reward_points",
            "interest_rate",
            "fee_waiver",
            "miles_earned",
            "category",
            "valid_from",
            "valid_to",
            "status"
          ],
          "title": "OfferCreate"
        },
        "OfferResponse": {
          "properties": {
            "offer_name": {
              "type": "string",
              "title": "Offer Name"
            },
            "offer_type": {
              "type": "string",
              "title": "Offer Type"
            },
            "cashback_percentage": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cashback Percentage"
            },
            "discount_amount": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Discount Amount"
            },
            "reward_points": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Reward Points"
            },
            "interest_rate": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Interest Rate"
            },
            "fee_waiver": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Fee Waiver"
            },
            "miles_earned": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Miles Earned"
            },
            "category": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            },
            "valid_from": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Valid From"
            },
            "valid_to": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Valid To"
            },
            "status": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            },
            "offer_id": {
              "type": "integer",
              "title": "Offer Id"
            }
          },
          "type": "object",
          "required": [
            "offer_name",
            "offer_type",
            "cashback_percentage",
            "discount_amount",
            "reward_points",
            "interest_rate",
            "fee_waiver",
            "miles_earned",
            "category",
            "valid_from",
            "valid_to",
            "status",
            "offer_id"
          ],
          "title": "OfferResponse"
        },
...
        "ValidationError": {
          "properties": {
            "loc": {
              "items": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "integer"
                  }
                ]
              },
              "type": "array",
              "title": "Location"
            },
            "msg": {
              "type": "string",
              "title": "Message"
            },
            "type": {
              "type": "string",
              "title": "Error Type"
            }
          },
          "type": "object",
          "required": [
            "loc",
            "msg",
            "type"
          ],
          "title": "ValidationError"
        }
      }
    }
  }